+++ examples/example-next-13/src/app/[locale]/page.tsx
@@ -1,16 +1,21 @@
 'use client';
 
 import {useTranslations} from 'next-intl';
-import LocaleSwitcher from '../../components/LocaleSwitcher';
-import PageLayout from '../../components/PageLayout';
+import PageLayout from 'components/PageLayout';
 
-export default function Index() {
-  const t = useTranslations('Index');
+export default function IndexPage() {
+  const t = useTranslations('IndexPage');
 
   return (
     <PageLayout title={t('title')}>
-      <p>{t('description')}</p>
-      <LocaleSwitcher />
+      <p className="max-w-[590px]">
+        {t.rich('description', {
+          p: (chunks) => <p className="mt-4">{chunks}</p>,
+          code: (chunks) => (
+            <code className="font-mono text-white">{chunks}</code>
+          )
+        })}
+      </p>
     </PageLayout>
   );
 }