+++ packages/use-intl/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 2.14.1 (2023-05-11)
+
+
+### Bug Fixes
+
+* Fix support for older Next.js versions by moving `Link` to `next-intl/link` ([#288](https://github.com/amannn/next-intl/issues/288)) ([f26ef99](https://github.com/amannn/next-intl/commit/f26ef999bf92c142d56d0009259e5a224c5dec5b)), closes [#287](https://github.com/amannn/next-intl/issues/287)
+
+
+
+
+
+# 2.14.0 (2023-05-10)
+
+
+### Features
+
+* Add navigation APIs for App Router (`useRouter`, `usePathname` and `Link`) ([#282](https://github.com/amannn/next-intl/issues/282)) ([e30a89b](https://github.com/amannn/next-intl/commit/e30a89b7079d31cfdefdd1a2d0c0a750adf3a6ce))
+
+
+
+
+
+## 2.13.4 (2023-05-05)
+
+**Note:** Version bump only for package use-intl
+
+
+
+
+
+## 2.13.3 (2023-05-05)
+
+**Note:** Version bump only for package use-intl
+
+
+
+
+
 ## 2.13.2 (2023-05-03)